Moved channel ID conversion into helper.

+++ TrendingGiphyBot/TrendingGiphyBotFunctions/Helpers/IDiscordHelper.cs
@@ -6,7 +6,7 @@ namespace TrendingGiphyBotFunctions.Helpers
     //TODO rename this to Wrapper
     public interface IDiscordHelper
     {
-        Task<IMessageChannel> GetChannelAsync(ulong id);
+        Task<IMessageChannel> GetChannelAsync(decimal id);
         Task LogInAsync();
         Task LogOutAsync();
     }
